droplet: implement access to volume by iterating
Before dpl_opendir has been used to access all chunks of a volume.
Unfortenatly, it has a bug and returns only the first 1000 entries.

As we know, that chunks are always named as 4 digit numbers from 0000 to 9999
we now iterate through them. We stop, if a chunk does not exist.

This commit also fixes the check_remote function.
Before, if droplet have already be initialized
but connection to backend stops operating,
check_remote still return true. This is now fixed.